Refactor utilities, extract common file operations
